created timestream module and split class up into many files
