+++ .vscode/settings.json
@@ -1,4 +1,13 @@
 {
     "rust-analyzer.cargo.features": "all",
     "editor.formatOnSave": true,
+    "cSpell.words": [
+        "fieldref",
+        "nanos",
+        "powf",
+        "pyarrow",
+        "subsort",
+        "Timestream",
+        "Timestreams"
+    ],
 }
