+++ python/docs/source/reference/execution.md
@@ -8,4 +8,5 @@
 
     Execution
     ResultIterator
+    record
 ```
