+++ python/docs/source/reference/timestream/records.md
@@ -11,9 +11,8 @@ Record operations create, extract or manipulate Timestreams of records.
    :toctree: ../apidocs/
 
     Timestream.col
-    Timestream.select
-    Timestream.remove
     Timestream.extend
     Timestream.record
-    record
+    Timestream.remove
+    Timestream.select
 ```
